packages/requirements.txt
jinja2
bs4
requests
tzlocal
Cheetah3